// app/routes/app.settings.modules.new/form-instruction/schema.ts
import { z } from 'zod'
import { zfd } from 'zod-form-data'

export const InstructionSchema = z.object({
	title: z.string(),
	answerKey: z.string().nullable().optional(),
	answerType: z.string(),
	answerTypeOptions: z.string().nullable().optional(),
	prompt: z.string(),
	promptType: z.string(),
	position: zfd.numeric(z.number().min(0)).optional(),
	concludingPrompt: z.string().optional(),
	concludingPromptType: z.string().optional(),
	canAskQuestion: z.union([
		z.literal('true').transform(() => true),
		z.literal('false').transform(() => false),
	]),
})